tools/requirements.txt
Pillow
pydub
opencv-python
numpy
